zaino-primitives: split zatoshis into a module directory
Promote the flat `types/zatoshis.rs` into a module root plus a
`types/zatoshis/` directory. `Zatoshis` and `ZatoshisOverflow` move into
`amount.rs`; `SignedZatoshis` and `SignedZatoshisOverflow` move into
`delta.rs`. Both moves are verbatim.

`MAX_ZATOSHIS` sits in `amount.rs` beside the amount it bounds and is
reachable from the sibling modules through the module root. The public
re-export surface in `types.rs` is unchanged.

Prepares the module for the quantity-arithmetic family in ADR-0013.
